Move kubernetes client into it's own package
