+++ pkg/skaffold/kubernetes/watcher_test.go
@@ -28,6 +28,7 @@ import (
 	"k8s.io/client-go/kubernetes/fake"
 	k8stesting "k8s.io/client-go/testing"
 
+	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/kubernetes/client"
 	"github.com/GoogleContainerTools/skaffold/testutil"
 )
 
@@ -66,7 +67,7 @@ func TestPodWatcher(t *testing.T) {
 	})
 
 	testutil.Run(t, "fail to get client", func(t *testutil.T) {
-		t.Override(&Client, func() (kubernetes.Interface, error) { return nil, errors.New("unable to get client") })
+		t.Override(&client.Client, func() (kubernetes.Interface, error) { return nil, errors.New("unable to get client") })
 
 		watcher := NewPodWatcher(&anyPod{}, []string{"ns"})
 		watcher.Register(make(chan PodEvent))
@@ -78,7 +79,7 @@ func TestPodWatcher(t *testing.T) {
 
 	testutil.Run(t, "fail to watch pods", func(t *testutil.T) {
 		clientset := fake.NewSimpleClientset()
-		t.Override(&Client, func() (kubernetes.Interface, error) { return clientset, nil })
+		t.Override(&client.Client, func() (kubernetes.Interface, error) { return clientset, nil })
 
 		clientset.Fake.PrependWatchReactor("pods", func(action k8stesting.Action) (handled bool, ret watch.Interface, err error) {
 			return true, nil, errors.New("unable to watch")
@@ -94,7 +95,7 @@ func TestPodWatcher(t *testing.T) {
 
 	testutil.Run(t, "filter 3 events", func(t *testutil.T) {
 		clientset := fake.NewSimpleClientset()
-		t.Override(&Client, func() (kubernetes.Interface, error) { return clientset, nil })
+		t.Override(&client.Client, func() (kubernetes.Interface, error) { return clientset, nil })
 
 		podSelector := &hasName{
 			validNames: []string{"pod1", "pod2", "pod3"},
